+++ runtime/tascam/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
 edition = "2018"
-name = "tascam-runtime"
+name = "snd-firewire-tascam-ctl-service"
 version = "0.1.0"
 # For publishing.
 authors = ["Takashi Sakamoto"]
 description = """
-Implementaion of service runtime for models of Tascam FireWire series
+Runtime implementation to control audio and music unit in IEEE 1394 bus which ALSA firewire-tascam
+(snd-firewire-tascam) supports as well as FE-8.
 """
 license = "GPL-3.0-or-later"
 publish = false
