+++ runtime/tascam/src/main.rs
@@ -16,7 +16,8 @@ mod seq_cntr;
 use {
     alsaseq::{prelude::*, *},
     asynch_runtime::*,
-    core::{card_cntr::*, *},
+    clap::Parser,
+    core::{card_cntr::*, cmdline::*, LogLevel, *},
     firewire_tascam_protocols as protocols,
     glib::{source, Error, FileError, IsA},
     hinawa::{
@@ -216,4 +217,29 @@ where
             })
         })
     }
+}
+
+struct TascamServiceCmd;
+
+#[derive(Parser, Default)]
+#[clap(name = "snd-firewire-tascam-ctl-service")]
+struct Arguments {
+    /// The name of subsystem; 'snd' or 'fw'.
+    subsystem: String,
+    /// The numeric identifier of sound card or firewire character device.
+    sysnum: u32,
+
+    /// The level to debug runtime, disabled as a default.
+    #[clap(long, short, arg_enum)]
+    log_level: Option<LogLevel>,
+}
+
+impl ServiceCmd<Arguments, (String, u32), TascamRuntime> for TascamServiceCmd {
+    fn params(args: &Arguments) -> ((String, u32), Option<LogLevel>) {
+        ((args.subsystem.clone(), args.sysnum), args.log_level)
+    }
+}
+
+fn main() {
+    TascamServiceCmd::run()
 }
